Add files via upload
Fixed the error with the hasPosition, it is now a data property. I also create selfEmployed and selfEmployedBy and separated them from the Employs and EmployedBy. I also added in new examples that verify that I am able to make more inferences (i.e. that a person that is asserted to be EmployedBy an Organization is inferred as an employee)